addressed PR feedbacks
